create helper file for fetch funcs, add states for adding new data
